requirements.txt
langchain
langchain-ollama
ollama
watchdog
pyyaml
openai-whisper
obs-websocket-py
pydrive2
